requirements.txt
Django==3.1.7
django-discordbot==1.1.7